Add interface class and second OI class
